[front] Create Paper internal component
