+++ openex-front/src/components/common/Paper.tsx
@@ -1,0 +1,29 @@
+import { makeStyles } from '@mui/styles';
+import PaperMui from '@mui/material/Paper';
+import React, { FunctionComponent } from 'react';
+
+interface PaperProps {
+  children: React.ReactElement
+}
+
+const useStyles = makeStyles((theme) => ({
+  paper: {
+    padding: 20,
+    marginBottom: 30,
+    borderRadius: 6,
+  },
+}));
+
+const Paper: FunctionComponent<PaperProps> = ({
+  children
+}) => {
+  const classes = useStyles();
+
+  return (
+    <PaperMui variant="outlined" className={classes.paper}>
+      {children}
+    </PaperMui>
+  )
+}
+
+export default Paper;
