+++ openex-front/tsconfig.json
@@ -18,9 +18,8 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx"
+    "jsx": "react-jsx",
+    "baseUrl": ".",
   },
-  "include": [
-    "src"
-  ]
+  "include": ["src"]
 }
